Update index.html
Mmmmmmmmm, I'm a dumbasss.... should you be making a 'movie' app in your code? Or maybe I should be in charge of that and can keep it in the public folder... Like, right now, you have this nice CSS set-up. Would maybe the next step be to make an object for Rush Hour with picture, title, and description, and then I'll put it into this section?